Hide moz-extension:// URL from page javascript
The moz-extension:// URL is a random UUID specific to the user
profile, i.e. it directly identifies a user and thus should not be
visible to page javascript. According to Mozilla bug 1372288, reads of
window.location should be protected by same-origin policy, so using
this approach should hide the URL from page javascript. This mitigates
the privacy aspect of issue #7, but as the iframe still has CSS rules
and a (static) UUID as an id, there is still stuff left that can be
used for fingerprinting, but now it only gives "a user profile with
LSR installed" instead of "this specific user".