Set overlay iframe src to a dummy value.
Some time ago I came across a site that, when Link Status Redux and
javascript were enabled for that site, exhibited serious breakage: 404
error pages for existing pages, weird layer and frame-within-frame
layout. After some investigation I found the culprit: the site used a
toolkit/framework that appends the string "?wmode=transparent" to all
iframe src attributes via javascript. (Later, this exact same behavior
with the exact same string was confirmed by user Sumer123 at AMO for a
different site, so looks like it's definitely a toolkit/framework.)

The problem is that commit fd59dca leaves the iframe src unset for
privacy reasons discussed in issue #7. When the page javascript does
the appending, the src now becomes "undefined?wmode=transparent",
which is a valid URL relative to the current directory on the web
server, which in turn leads to a 404 page or similar being the content
of the topmost, fullscreen overlay iframe.

The solution is to have a value for the src attribute, one which is
not unique to every user, and one that does not actually change the
content of the iframe when something is appended to it. The latter
property is the challenge, but the chosen "moz-extension://none/"
seems to work, at least for the two problem sites, and should be safe
for appending. (I tried "null://", as I would have preferred page
javascript to not be trivially able to count extension iframes, but it
did not work.)

All this demonstrates once again that the WebExtension API is
completely inadequate for creating UI elements. This fix works for
appends, but nothing prevents page javascript from manipulating the
iframe src in other ways, which may lead to similar breakage, and
there is nothing (well, nothing that is not subject to race
conditions) that a WebExtension can do about that.