Compensate for horizontal scrollbar height in popup bottom offset.
When the overlink popup is shown above the browser's native one, there
is an unintentional gap between the two if the page has a horizontal
scrollbar, as the bottom offset of the add-on popup is relative to the
scrollbar's top edge. This patch finds out and subtracts the scrollbar
height from the bottom offset, eliminating the gap.

Fixes #13.