Fix horizontal scrollbar height compensation for frames.
The horizontal scrollbar height compensation added in commit c3a61e0
did not work for links inside (i)frames. This commit fixes it by
sending the initial window height (and listening for resize events) of
the top window from the content script, and keeping track in the
background script of each tab's window height that can be sent in the
"show" message to the overlay.

Fixes #13.