Compensate for browser zoom level
Listen for browser.tabs.onZoomChange events in the background script,
and send the tab zoom levels to the overlay. In the overlay, scale the
popup panel with the inverse of the zoom level, and also use the zoom
level to convert the mouse coordinates (which are HiDPI-scaled but NOT
zoom-scaled) into CSS pixels to achieve correct positioning when the
popup panel is configured to be at the mouse pointer.

The popup panel scaling is achieved via CSS transform. As the
transform seems to be applied after all other CSS rules (at least
conceptually) the result is not perfect (possible ugliness in border
widths and font kerning and antialiasing) but it is mostly more
desirable than without scaling.

Fixes #12; fixes #23.