Add indicator for links that are currently open in a tab
This commit adds only the functionality of indicating currently open
links (enabled by default), no option UI changes are included. A
currently-open indicator prefix overrides a visited prefix (as currently
open implies visited), and is prepended to a bookmarked prefix
(bookmarked implies visited but not currently open).

This feature has a known problem with false negatives: HTTP redirections
and URI fragments may make a link seem to be not currently open even
though it (or rather, its final destination) is. The option UI will need
a prominent note about this.