Prevent reset of the featureLevel preference
options.js has a cleanup routine where preferences with the default
value *and* those that are not present as an input field are removed
from the stored preferences. This means that featureLevel needs to be
added as a hidden input field, otherwise it will be removed from the
preferences when the settings page is opened, causing the news page to
display spuriously on next extension startup.