Workaround for issue #41
With Firefox 102.0b3 using LSR with dark theme on certain pages (I'm
guessing those that have certain kind of adjustments for light/dark
mode) makes the whole page white. The culprit is a (pseudo-)layer
between the overlay iframe and the overlay document that is white with
dark theme but (properly) transparent with light theme. This seems to
be a Firefox bug as it affects only the dark theme(s) and it was not
present in earlier versions.

The color/transparency of the layer cannot be directly addressed in
CSS, but setting the property "color-scheme: light dark" for both the
iframe *and* the overlay document root seems to fix the problem.