Fix feature level finalization
When saving a feature level value, save it negated so that it won't
get deleted from storage by the preferences housekeeping code due to
being equal to the default value. The alternative would be to get rid
of the housekeeping code in question, but I think it could still be
useful, so keep it for now.