Better workaround for issue #41
Change the color-scheme combination so that the workaround works even
when the iframe content is not (yet) loaded, in which case the empty
content will have the default "normal" color-scheme. This should fix
white flickering with dark theme when (re)loading a page, and hopefully
also compatibility issues with web pages and other addons (such as
SingleFile mentioned in the issue).