Fix clipping of the URL bottom (hopefully)
The original line and div height of 1.2em was chosen because the MDN
documentation for line-height says that the default is roughly
1.2. That is a unitless number, which we cannot use because we need to
specify div heights, but using 1.2em seemed to work just fine for this
purpose. Almost. Depending on font/OS(/other variables?), 0-2 CSS
pixels of the descenders (e.g. the lower hook of "g") are clipped by
the overflow:hidden setting of the URL div.

What we are *trying* to do here is to have a div with the same height
as the full font height, but AFAICT the latter measure is not
available to CSS, so we must resort to using em's. (Also, setting the
line height to the same value as the div height we can have multi-line
divs with exactly N lines visible.) Increasing the line and div height
to 1.3em fixes the clipping when tested on Linux and Windows 10/11
using 1x/1.5x/2x DPI scaling. This fix is simple but not robust, as it
may not work for all fonts, but I cannot think of a "proper" way to
fix this, at least not without making it a lot more complex.

Fixes #44.