Remove web_accessible_resources
Apparently this declaration for the overlay documents is needed on
Chrome, but not on Firefox which has privileged content scripts. I
have no intention/desire/time to support Chrome, so remove it for
theoretical improvement of security and privacy. In practice the
improvement is insignificant, as 1) Firefox uses randomized extension
URLs and 2) the documents themselves do not contain sensitive
information, the sensitive information being in the live state of the
overlay which is not accessible to the "host" page.