Inject overlay on-demand on first link hover
This is a partial workaround to reduce site breakage in cases where
the page is (re)built by JS after loading and is checked for
"tampering" in the process (e.g. when using React; see issue #47). The
caveat is that this will not work if the mouse is already over a link
on page load, or is moved over a link too soon after page load.